version 1.1.1.11
git-svn-id: http://svn.openvpn.net/projects/openvpn/trunk/openvpn@22 e7ae566f-a301-0410-adde-c780ea21d3b5

+++ buffer.c
@@ -194,8 +194,8 @@ debug_gc_check_corrupt (const char *file, int line)
   while (e = stack)
     {
       if (e->level > thread->gc_level)
-	printf ("GC CORRUPT 0x%08zx lev=%d back=0x%08zx file=%s line=%d\n", e,
-		e->level, e->back, file, line);
+	printf ("GC CORRUPT 0x%08zx lev=%d back=0x%08zx file=%s line=%d\n",
+		(size_t)e, e->level, (size_t)e->back, file, line);
       stack = e->back;
     }
 }
